Homepage/menu changes, partial course listing

+++ laravel/app/database/migrations/2014_12_09_205628_add_original_affiliator_to_users_table.php
@@ -14,7 +14,7 @@ class AddOriginalAffiliatorToUsersTable extends Migration {
 	{
 		Schema::table('users', function(Blueprint $table)
 		{
-			$table->integer('ltc_affiliator_id')->after('id');
+			$table->bigInteger('ltc_affiliator_id')->after('id');
 		});
 	}
